Gradle on Docker steroids.
Both `run-tests.sh` and `run-task.sh` go out of their way to ensure
that the processes being run on your behalf inside of the docker
container are being invoked with the same UID:GID as the current user
invoking the scripts. Thus ensuring that any files being created will
have the same UID:GID as the current user. Is this a massive hack? You
be the judge!